feat: Fetch and watch for Cilium policies stored in kvstore
This change adds a feature that allows Cilium to watch a key-prefix in a support kvstore for Cilium network policies. It's modelled on the work done in https://github.com/cilium/cilium/pull/32599 which added support for the same feature but from a filesystem.

The reason to use a kvstore backend is for non-Kubernetes environments that are running containers. They still need dynamic policy updates but cannot use CRDs. The directory watcher can be made to work but is much easier with a central store. It would also be possible to use the gRPC API to do this, however the recent policy repository optimisations make it easier to do it within the Cilium agent due to the indexing by resource ID.